Implementação de validações e notificações finalizada

+++ ApiTresCamadas/src/CKLabs.Business/Services/FornecedorService.cs
@@ -1,5 +1,7 @@
-﻿using CKLabs.Business.Interfaces;
+﻿using CKLabs.Business.Helpers;
+using CKLabs.Business.Interfaces;
 using CKLabs.Business.Models;
+using CKLabs.Business.Models.Validations;
 
 namespace CKLabs.Business.Services
 {
@@ -16,7 +18,7 @@ namespace CKLabs.Business.Services
         /// <summary>
         /// Construtor
         /// </summary>
-        public FornecedorService(IFornecedorRepository fornecedorRepository)
+        public FornecedorService(IFornecedorRepository fornecedorRepository, INotificador notificador) : base(notificador)
         {
             _fornecedorRepository = fornecedorRepository;
         }
@@ -28,8 +30,14 @@ namespace CKLabs.Business.Services
         /// <returns></returns>
         public async Task Adicionar(Fornecedor fornecedor)
         {
-            // Validar se a entidade é consistente...
-            // Validar se já existe um fornecedor com o mesmo documento cadastrado...
+            if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)
+                || !ExecutarValidacao(new EnderecoValidation(), fornecedor.Endereco)) return;
+
+            if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento).Result.Any())
+            {
+                Notificar(ValidationMessages.JaExisteFornecedorComDocumentoInformado());
+                return;
+            }
 
             await _fornecedorRepository.Adicionar(fornecedor);
         }
@@ -41,6 +49,14 @@ namespace CKLabs.Business.Services
         /// <returns></returns>
         public async Task Atualizar(Fornecedor fornecedor)
         {
+            if (!ExecutarValidacao(new FornecedorValidation(), fornecedor)) return;
+
+            if (_fornecedorRepository.Buscar(f => f.Documento == fornecedor.Documento && f.Id != fornecedor.Id).Result.Any())
+            {
+                Notificar(ValidationMessages.JaExisteFornecedorComDocumentoInformado());
+                return;
+            }
+
             await _fornecedorRepository.Atualizar(fornecedor);
         }
 
@@ -51,6 +67,27 @@ namespace CKLabs.Business.Services
         /// <returns></returns>
         public async Task Remover(Guid id)
         {
+            var fornecedor = await _fornecedorRepository.ObterFornecedorProdutosEndereco(id);
+
+            if (fornecedor == null)
+            {
+                Notificar("Fornecedor não encontrado");
+                return;
+            }
+
+            if (fornecedor.Produtos.Any())
+            {
+                Notificar("O fornecedor possui produtos cadastrados");
+                return;
+            }
+
+            var endereco = await _fornecedorRepository.ObterEnderecoPorFornecedor(id);
+
+            if (endereco != null)
+            {
+                await _fornecedorRepository.RemoverEnderecoFornecedor(endereco);
+            }
+
             await _fornecedorRepository.Remover(id);
         }
